Move Gradle cache exclude paths related functionality to gradle.go file.
